+++ .github/actions/test/action.yml
@@ -10,14 +10,14 @@ inputs:
     description: Token to log into docker hub
   with-integration-tests:
     required: false
-    default: 'true'
+    default: "true"
     description: If enabled, runs integration tests and needs docker
   datadog-api-key:
     required: true
     description: API key for datadog uploads
 
 runs:
-  using: 'composite'
+  using: "composite"
   steps:
     - uses: oven-sh/setup-bun@v1
       if: runner.os != 'Windows'
@@ -27,7 +27,7 @@ runs:
     - name: Build wrappers
       shell: bash
       if: runner.os != 'Windows'
-      working-directory: './cli/wrappers'
+      working-directory: "./cli/wrappers"
       run: |
         bun i
         bun run build
@@ -35,7 +35,7 @@ runs:
     - name: Build wrappers
       shell: bash
       if: runner.os == 'Windows'
-      working-directory: './cli/wrappers'
+      working-directory: "./cli/wrappers"
       run: |
         powershell -c "irm bun.sh/install.ps1 | iex"
         ~/.bun/bin/bun.exe i
@@ -79,7 +79,7 @@ runs:
       shell: bash
       run: |
         docker compose -f engine/crates/integration-tests/docker-compose.yml up -d
-        RUST_BACKTRACE=1 cargo nextest run --workspace --profile ci --exclude grafbase-gateway --exclude grafbase-docker-tests
+        RUST_BACKTRACE=1 cargo nextest run --workspace --profile ci --exclude gateway-integration-tests --exclude grafbase-docker-tests
         docker compose -f engine/crates/integration-tests/docker-compose.yml stop -t 3
 
     - name: Upload the JUnit files
@@ -95,8 +95,8 @@ runs:
       if: ${{ inputs.with-integration-tests == 'true' }}
       shell: bash
       run: |
-        docker compose -f gateway/crates/gateway-binary/docker-compose.yml up -d
-        RUST_BACKTRACE=1 cargo nextest run -p grafbase-gateway --profile ci
+        docker compose -f gateway/crates/integration-tests/docker-compose.yml up -d
+        RUST_BACKTRACE=1 cargo nextest run -p gateway-integration-tests --profile ci
 
     - name: Upload the gateway junit files
       if: ${{ inputs.with-integration-tests == 'true' && ( success() || failure() ) && !contains(steps.gateway_tests.outputs.exitcode, '101') }}
@@ -111,7 +111,7 @@ runs:
       if: ${{ inputs.with-integration-tests == 'false' }}
       shell: bash
       run: |
-        RUST_BACKTRACE=1 cargo nextest run --workspace --exclude integration-tests --exclude grafbase-docker-tests --exclude grafbase-gateway --exclude wasi-component-loader --profile ci
+        RUST_BACKTRACE=1 cargo nextest run --workspace --exclude integration-tests --exclude grafbase-docker-tests --exclude gateway-integration-tests --exclude wasi-component-loader --profile ci
 
     - name: Upload the non-integration JUnit files
       if: ${{ inputs.with-integration-tests == 'false' && ( success() || failure() ) && !contains(steps.tests_no_integration.outputs.exitcode, '101') }}
