--- gateway/crates/gateway-binary/tests/mocks/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod gdn;
