+++ gateway/crates/integration-tests/tests/integration_tests.rs
@@ -1,27 +1,17 @@
-#![allow(unused_crate_dependencies, clippy::panic)]
+#![allow(unused_crate_dependencies)]
 
-mod mocks;
 mod telemetry;
 
-use std::{
-    borrow::Cow,
-    env, fs,
-    marker::PhantomData,
-    net::{Ipv4Addr, SocketAddr, SocketAddrV4},
-    panic::{catch_unwind, AssertUnwindSafe},
-    path,
-    sync::{Arc, Mutex, OnceLock},
-    time::{Duration, SystemTime},
-};
+use std::{fs, net::SocketAddr, panic::AssertUnwindSafe, sync::Arc, time::Duration};
 
-use crate::mocks::gdn::GdnResponseMock;
-use duct::{cmd, Handle};
-use futures_util::future::BoxFuture;
+use duct::cmd;
 use futures_util::{Future, FutureExt};
-use http::{HeaderMap, StatusCode};
+use gateway_integration_tests::mocks::gdn::GdnResponseMock;
+use gateway_integration_tests::{
+    cargo_bin, listen_address, runtime, Client, CommandHandles, ConfigContent, GatewayBuilder, TestRequest,
+};
 use indoc::indoc;
 use tempfile::tempdir;
-use tokio::runtime::Runtime;
 use tokio::time::Instant;
 use wiremock::{
     matchers::{header, method, path},
@@ -29,417 +19,6 @@ use wiremock::{
 };
 
 const ACCESS_TOKEN: &str = "test";
-
-#[derive(serde::Serialize)]
-struct BatchQuery {
-    query: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    variables: Option<serde_json::Value>,
-}
-
-#[must_use]
-pub struct GqlBatchBuilder<Response> {
-    queries: Vec<BatchQuery>,
-    phantom: PhantomData<fn() -> Response>,
-    reqwest_builder: reqwest::RequestBuilder,
-    bearer: Option<String>,
-}
-
-impl<Response> GqlBatchBuilder<Response> {
-    pub async fn send(self) -> Response
-    where
-        Response: for<'de> serde::de::Deserialize<'de>,
-    {
-        let json = serde_json::to_value(&self.queries).expect("to be able to serialize gql request");
-
-        if let Some(bearer) = self.bearer {
-            self.reqwest_builder.header("authorization", bearer)
-        } else {
-            self.reqwest_builder
-        }
-        .json(&json)
-        .send()
-        .await
-        .unwrap()
-        .json::<Response>()
-        .await
-        .unwrap()
-    }
-}
-
-#[derive(serde::Serialize)]
-#[must_use]
-pub struct GqlRequestBuilder<Response> {
-    // These two will be serialized into the request
-    query: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    variables: Option<serde_json::Value>,
-
-    // These won't
-    #[serde(skip)]
-    phantom: PhantomData<fn() -> Response>,
-    #[serde(skip)]
-    reqwest_builder: reqwest::RequestBuilder,
-    #[serde(skip)]
-    bearer: Option<String>,
-}
-
-impl<Response> GqlRequestBuilder<Response> {
-    pub fn variables(mut self, variables: impl serde::Serialize) -> Self {
-        self.variables = Some(serde_json::to_value(variables).expect("to be able to serialize variables"));
-        self
-    }
-
-    pub fn bearer(mut self, token: &str) -> Self {
-        self.bearer = Some(format!("Bearer {token}"));
-        self
-    }
-
-    pub fn header(self, name: &str, value: &str) -> Self {
-        let Self {
-            bearer,
-            phantom,
-            query,
-            mut reqwest_builder,
-            variables,
-        } = self;
-        reqwest_builder = reqwest_builder.header(name, value);
-        Self {
-            query,
-            variables,
-            phantom,
-            reqwest_builder,
-            bearer,
-        }
-    }
-
-    pub async fn send(self) -> Response
-    where
-        Response: for<'de> serde::de::Deserialize<'de>,
-    {
-        let json = serde_json::to_value(&self).expect("to be able to serialize gql request");
-
-        if let Some(bearer) = self.bearer {
-            self.reqwest_builder.header("authorization", bearer)
-        } else {
-            self.reqwest_builder
-        }
-        .json(&json)
-        .send()
-        .await
-        .unwrap()
-        .json::<Response>()
-        .await
-        .unwrap()
-    }
-
-    pub async fn request(self) -> reqwest::Response {
-        let json = serde_json::to_value(&self).expect("to be able to serialize gql request");
-        self.reqwest_builder.json(&json).send().await.unwrap()
-    }
-}
-
-pub struct Client {
-    endpoint: String,
-    client: reqwest::Client,
-    headers: HeaderMap,
-    commands: CommandHandles,
-}
-
-impl Client {
-    pub fn new(endpoint: String, commands: CommandHandles) -> Self {
-        Self {
-            endpoint,
-            headers: HeaderMap::new(),
-            client: reqwest::Client::builder()
-                .connect_timeout(Duration::from_secs(1))
-                .build()
-                .unwrap(),
-            commands,
-        }
-    }
-
-    pub fn with_header(mut self, key: &'static str, value: impl AsRef<str>) -> Self {
-        self.headers.insert(key, value.as_ref().parse().unwrap());
-        self
-    }
-
-    pub async fn poll_endpoint(&self, timeout_secs: u64, interval_millis: u64) {
-        let start = SystemTime::now();
-
-        loop {
-            let valid_response = self
-                .client
-                .head(&self.endpoint)
-                .send()
-                .await
-                .is_ok_and(|response| response.status() != StatusCode::SERVICE_UNAVAILABLE);
-
-            if valid_response {
-                break;
-            }
-
-            assert!(start.elapsed().unwrap().as_secs() < timeout_secs, "timeout");
-
-            tokio::time::sleep(Duration::from_millis(interval_millis)).await;
-        }
-    }
-
-    pub fn kill_handles(&self) {
-        self.commands.kill_all()
-    }
-
-    pub fn endpoint(&self) -> &str {
-        &self.endpoint
-    }
-
-    pub fn gql<Response>(&self, query: impl Into<String>) -> GqlRequestBuilder<Response>
-    where
-        Response: for<'de> serde::de::Deserialize<'de>,
-    {
-        let reqwest_builder = self.client.post(&self.endpoint).headers(self.headers.clone());
-
-        GqlRequestBuilder {
-            query: query.into(),
-            variables: None,
-            phantom: PhantomData,
-            reqwest_builder: reqwest_builder.header(http::header::ACCEPT, "application/json"),
-            bearer: None,
-        }
-    }
-
-    pub fn gql_batch<Response, T>(&self, queries: impl IntoIterator<Item = T>) -> GqlBatchBuilder<Response>
-    where
-        Response: for<'de> serde::de::Deserialize<'de>,
-        T: Into<String>,
-    {
-        let reqwest_builder = self.client.post(&self.endpoint).headers(self.headers.clone());
-
-        let queries = queries
-            .into_iter()
-            .map(|query| BatchQuery {
-                query: query.into(),
-                variables: None,
-            })
-            .collect();
-
-        GqlBatchBuilder {
-            queries,
-            phantom: PhantomData,
-            reqwest_builder: reqwest_builder.header(http::header::ACCEPT, "application/json"),
-            bearer: None,
-        }
-    }
-
-    pub fn client(&self) -> &reqwest::Client {
-        &self.client
-    }
-}
-
-#[derive(Clone)]
-pub struct CommandHandles(Arc<Mutex<Vec<Handle>>>);
-
-impl CommandHandles {
-    pub fn new() -> Self {
-        CommandHandles(Arc::new(Mutex::new(vec![])))
-    }
-
-    pub fn push(&mut self, handle: Handle) {
-        self.0.lock().unwrap().push(handle);
-    }
-
-    pub fn still_running(&self) -> bool {
-        self.0
-            .lock()
-            .unwrap()
-            .iter()
-            .all(|handle| handle.try_wait().unwrap().is_none())
-    }
-
-    pub fn kill_all(&self) {
-        for command in self.0.lock().unwrap().iter() {
-            command.kill().unwrap();
-        }
-    }
-}
-
-impl Default for CommandHandles {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-pub fn cargo_bin<S: AsRef<str>>(name: S) -> path::PathBuf {
-    cargo_bin_str(name.as_ref())
-}
-
-fn target_dir() -> path::PathBuf {
-    env::current_exe()
-        .ok()
-        .map(|mut path| {
-            path.pop();
-            if path.ends_with("deps") {
-                path.pop();
-            }
-            path
-        })
-        .unwrap()
-}
-
-fn cargo_bin_str(name: &str) -> path::PathBuf {
-    let env_var = format!("CARGO_BIN_EXE_{name}");
-    std::env::var_os(env_var).map_or_else(
-        || target_dir().join(format!("{name}{}", env::consts::EXE_SUFFIX)),
-        std::convert::Into::into,
-    )
-}
-
-fn runtime() -> &'static Runtime {
-    static RUNTIME: OnceLock<Runtime> = OnceLock::new();
-    RUNTIME.get_or_init(|| {
-        tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .unwrap()
-    })
-}
-
-fn get_free_port() -> u16 {
-    const INITIAL_PORT: u16 = 14712;
-
-    let test_state_directory_path = std::env::temp_dir().join("grafbase/cli-tests");
-    std::fs::create_dir_all(&test_state_directory_path).unwrap();
-    let lock_file_path = test_state_directory_path.join("port-number.lock");
-    let port_number_file_path = test_state_directory_path.join("port-number.txt");
-    let mut lock_file = fslock::LockFile::open(&lock_file_path).unwrap();
-    lock_file.lock().unwrap();
-    let port_number = if port_number_file_path.exists() {
-        std::fs::read_to_string(&port_number_file_path)
-            .unwrap()
-            .trim()
-            .parse::<u16>()
-            .unwrap()
-            + 1
-    } else {
-        INITIAL_PORT
-    };
-    std::fs::write(&port_number_file_path, port_number.to_string()).unwrap();
-    lock_file.unlock().unwrap();
-    port_number
-}
-
-fn listen_address() -> SocketAddr {
-    let port = get_free_port();
-    SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), port))
-}
-
-struct ConfigContent<'a>(Option<Cow<'a, str>>);
-
-impl<'a> From<&'a str> for ConfigContent<'a> {
-    fn from(s: &'a str) -> Self {
-        ConfigContent(Some(Cow::Borrowed(s)))
-    }
-}
-
-impl<'a> From<&'a String> for ConfigContent<'a> {
-    fn from(s: &'a String) -> Self {
-        ConfigContent(Some(Cow::Borrowed(s)))
-    }
-}
-
-impl From<String> for ConfigContent<'static> {
-    fn from(s: String) -> Self {
-        ConfigContent(Some(Cow::Owned(s)))
-    }
-}
-
-struct GatewayBuilder<'a> {
-    toml_config: ConfigContent<'a>,
-    schema: &'a str,
-    log_level: Option<String>,
-    client_url_path: Option<&'a str>,
-    client_headers: Option<&'static [(&'static str, &'static str)]>,
-}
-
-impl<'a> GatewayBuilder<'a> {
-    fn new(schema: &'a str) -> Self {
-        Self {
-            toml_config: ConfigContent(None),
-            schema,
-            log_level: None,
-            client_url_path: None,
-            client_headers: None,
-        }
-    }
-
-    fn with_log_level(mut self, level: &str) -> Self {
-        self.log_level = Some(level.to_string());
-        self
-    }
-
-    fn run<F>(self, test: impl FnOnce(Arc<Client>) -> F)
-    where
-        F: Future<Output = ()>,
-    {
-        let temp_dir = tempdir().unwrap();
-        let config_path = temp_dir.path().join("grafbase.toml");
-
-        let schema_path = temp_dir.path().join("schema.graphql");
-        fs::write(&schema_path, self.schema).unwrap();
-
-        let addr = listen_address();
-        let mut args = vec![
-            "--listen-address".to_string(),
-            addr.to_string(),
-            "--schema".to_string(),
-            schema_path.to_str().unwrap().to_string(),
-        ];
-
-        if let Some(config) = self.toml_config.0 {
-            fs::write(&config_path, config.as_ref()).unwrap();
-            args.push("--config".to_string());
-            args.push(config_path.to_str().unwrap().to_string());
-        }
-
-        if let Some(level) = self.log_level {
-            args.push("--log".to_string());
-            args.push(level);
-        }
-
-        let command = cmd(cargo_bin("grafbase-gateway"), &args).stdout_null().stderr_null();
-
-        let endpoint = match self.client_url_path {
-            Some(path) => format!("http://{addr}/{path}"),
-            None => format!("http://{addr}/graphql"),
-        };
-
-        let mut commands = CommandHandles::new();
-        commands.push(command.start().unwrap());
-
-        let mut client = Client::new(endpoint, commands);
-
-        if let Some(headers) = self.client_headers {
-            for header in headers {
-                client = client.with_header(header.0, header.1);
-            }
-        }
-
-        let client = Arc::new(client);
-
-        let res = catch_unwind(AssertUnwindSafe(|| {
-            runtime().block_on(async {
-                client.poll_endpoint(30, 300).await;
-                test(client.clone()).await
-            })
-        }));
-
-        client.kill_handles();
-
-        if let Err(err) = res {
-            std::panic::resume_unwind(err);
-        }
-    }
-}
 
 fn with_static_server<'a, F, T>(
     config: impl Into<ConfigContent<'a>>,
@@ -530,21 +109,6 @@ async fn introspect(url: &str) -> String {
         .unwrap_or_default()
 }
 
-pub fn clickhouse_client() -> &'static ::clickhouse::Client {
-    static CLIENT: OnceLock<::clickhouse::Client> = OnceLock::new();
-    CLIENT.get_or_init(|| {
-        ::clickhouse::Client::default()
-            .with_url("http://localhost:8124")
-            .with_user("default")
-            .with_database("otel")
-    })
-}
-
-#[ctor::ctor]
-fn setup_rustls() {
-    rustls::crypto::ring::default_provider().install_default().unwrap();
-}
-
 // This failed when using `format_with()` inside tracing with opentelemetry. Somehow it gets called
 // multiple times which isn't supported and panics...
 #[test]
@@ -562,7 +126,7 @@ fn trace_log_level() {
     GatewayBuilder::new(&schema)
         .with_log_level("trace")
         .run(|client| async move {
-            let result: serde_json::Value = client.gql(query).send().await;
+            let result = client.execute(query).await.into_body();
             let result = serde_json::to_string_pretty(&result).unwrap();
 
             insta::assert_snapshot!(&result, @r###"
@@ -597,7 +161,7 @@ fn no_config() {
     "#};
 
     with_static_server(ConfigContent(None), &schema, None, None, |client| async move {
-        let result: serde_json::Value = client.gql(query).send().await;
+        let result = client.execute(query).await.into_body();
         let result = serde_json::to_string_pretty(&result).unwrap();
 
         insta::assert_snapshot!(&result, @r###"
@@ -632,7 +196,7 @@ fn static_schema() {
     "#};
 
     with_static_server("", &schema, None, None, |client| async move {
-        let result: serde_json::Value = client.gql(query).send().await;
+        let result = client.execute(query).await.into_body();
         let result = serde_json::to_string_pretty(&result).unwrap();
 
         insta::assert_snapshot!(&result, @r###"
@@ -754,7 +318,7 @@ fn custom_path() {
     "#};
 
     with_static_server(config, &schema, Some("custom"), None, |client| async move {
-        let result: serde_json::Value = client.gql(query).send().await;
+        let result = client.execute(query).await.into_body();
         let result = serde_json::to_string_pretty(&result).unwrap();
 
         insta::assert_snapshot!(&result, @r###"
@@ -794,8 +358,8 @@ fn csrf_no_header() {
     "#};
 
     with_static_server(config, &schema, Some("custom"), None, |client| async move {
-        let response = client.gql::<serde_json::Value>(query).request().await;
-        assert_eq!(http::StatusCode::FORBIDDEN, response.status());
+        let response = client.execute(query).await;
+        assert_eq!(http::StatusCode::FORBIDDEN, response.status);
     })
 }
 
@@ -819,7 +383,7 @@ fn csrf_with_header() {
     let headers = &[("x-grafbase-csrf-protection", "1")];
 
     with_static_server(config, &schema, None, Some(headers), |client| async move {
-        let result: serde_json::Value = client.gql(query).send().await;
+        let result = client.execute(query).await.into_body();
         let result = serde_json::to_string_pretty(&result).unwrap();
 
         insta::assert_snapshot!(&result, @r###"
@@ -854,7 +418,7 @@ fn hybrid_graph() {
     "#};
 
     with_hybrid_server("", "test_graph", &schema, |client, _, _| async move {
-        let result: serde_json::Value = client.gql(query).send().await;
+        let result = client.execute(query).await.into_body();
         let result = serde_json::to_string_pretty(&result).unwrap();
 
         insta::assert_snapshot!(&result, @r###"
@@ -1010,7 +574,7 @@ fn global_rate_limiting() {
     "#};
 
     with_static_server(config, &schema, None, None, |client| async move {
-        expect_rate_limiting(|| client.gql(query).send().boxed()).await;
+        expect_rate_limiting(|| client.execute(query)).await;
     })
 }
 
@@ -1033,7 +597,7 @@ fn subgraph_rate_limiting() {
     "#};
 
     with_static_server(config, &schema, None, None, |client| async move {
-        expect_rate_limiting(|| client.gql(query).send().boxed()).await;
+        expect_rate_limiting(|| client.execute(query)).await
     })
 }
 
@@ -1059,7 +623,7 @@ fn global_redis_rate_limiting() {
     "#};
 
     with_static_server(config, &schema, None, None, |client| async move {
-        expect_rate_limiting(|| client.gql(query).send().boxed()).await;
+        expect_rate_limiting(|| client.execute(query)).await
     })
 }
 
@@ -1085,20 +649,20 @@ fn subgraph_redis_rate_limiting() {
     "#};
 
     with_static_server(config, &schema, None, None, |client| async move {
-        expect_rate_limiting(|| client.gql(query).send().boxed()).await;
+        expect_rate_limiting(|| client.execute(query)).await
     })
 }
 
 #[allow(clippy::panic)]
-async fn expect_rate_limiting<'a, F>(f: F)
+async fn expect_rate_limiting<F>(f: F)
 where
-    F: Fn() -> BoxFuture<'a, serde_json::Value>,
+    F: Fn() -> TestRequest,
 {
     let destiny = Instant::now().checked_add(Duration::from_secs(60)).unwrap();
 
     loop {
-        let response = Box::pin(f());
-        let response = response.await;
+        let request = f();
+        let response = request.await.into_body();
 
         if response["errors"][0]["extensions"]["code"] == "RATE_LIMITED" {
             break;
