+++ gateway/crates/integration-tests/tests/schemas/small.graphql
